backend/pyproject.toml
[tool.poetry]
name = "alma-backend"
version = "0.1.0"
description = "Document upload and form auto-fill backend (BAML + FastAPI + Playwright)"
authors = []
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
python-multipart = "^0.0.12"
baml-py = "^0.73"
playwright = "^1.49.0"
httpx = "^0.27.0"
pydantic = "^2.0"
pymupdf = "^1.25.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "app.main:run"
